Move config dirs out of the properties service
